net/usrsock: replace usrsock xid type to uint64_t

+++ wireless/gs2200m/gs2200m_main.c
@@ -226,7 +226,7 @@ static int _write_to_usock(int fd, void *buf, size_t count)
  ****************************************************************************/
 
 static int _send_ack_common(int fd,
-                            uint8_t xid,
+                            uint64_t xid,
                             FAR struct usrsock_message_req_ack_s *resp)
 {
   resp->head.msgid = USRSOCK_MESSAGE_RESPONSE_ACK;
